fix build for gitee js 	modified:   BUILD.gn 	modified:   include/compiler.h 	modified:   src/compiler.cpp 	modified:   src/jsmirbuilder.cpp

+++ include/compiler.h
@@ -123,7 +123,7 @@ class JSCompiler {
   // Compile Functions.
   BaseNode *CompileOpConstValue(uint32_t jsvalue_tag, int32_t payload);
   js_builtin_id EcmaNameToId(char *name);
-  int32_t GetBuiltinStringId(const jschar *chars, uint32_t length);
+  int32_t GetBuiltinStringId(const jschar *chars, size_t length);
   BaseNode *CompileBuiltinObject(char *name);
   BaseNode *CompileBuiltinMethod(int32_t idx, int arg_num, bool need_this);
   uint32_t FindIntrinsicForOp(JSOp opcode);
